refactor: admin post routes

+++ app/Http/Controllers/Admin/MembersController.php
@@ -14,7 +14,7 @@ class MembersController extends Controller
 
     public function __construct()
     {
-        $this->middleware(['auth', 'auth.super']);
+        $this->middleware(['auth.super']);
     }
 
 
@@ -70,7 +70,7 @@ class MembersController extends Controller
             'linkedin' => $request->linkedin
         ]);
 
-        return redirect()->route('admin-members')->with('status', 'Member Added Successfully!');
+        return redirect()->route('admin.members')->with('status', 'Member Added Successfully!');
     }
 
     public function editExecutiveView(Request $request, $id) {
@@ -109,7 +109,8 @@ class MembersController extends Controller
         if ($request->hasFile('image')) {
             $prev_image = $member->image;
 
-            Storage::unlink('public/images/' . $prev_image);
+            if ($prev_image != "")
+                Storage::unlink('public/images/' . $prev_image);
 
             $image_name = time() . "-" . $request->name . "." . $request->image->extension();
 
@@ -124,7 +125,7 @@ class MembersController extends Controller
 
         $member->save();
 
-        return redirect()->route('admin-members')->with('status', 'Updated member successfully!');
+        return redirect()->route('admin.members')->with('status', 'Updated member successfully!');
     }
 
 
@@ -139,6 +140,6 @@ class MembersController extends Controller
         $member->delete();
 
         
-        return redirect()->route('admin-members')->with('status', 'Deleted member successfully!');
+        return redirect()->route('admin.members')->with('status', 'Deleted member successfully!');
     }
 }
